\name{fitCellGrowth}
\alias{fitCellGrowth}
\title{Fit growth curves}
\usage{
  fitCellGrowth(x, z, model = "locfit",
    locfit.h = 3 * 60 * 60, locfit.deg = 2,
    relative.height.at.lag = 0.1)
}
\arguments{
  \item{x}{\code{numeric} vector: time points}

  \item{z}{\code{numeric} vector: log(OD)}

  \item{model}{which model to fit.}

  \item{locfit.h}{\code{numeric}: \code{h} parameter
  (window size) in call to
  \code{\link[locfit:locfit]{locfit}}. The default value is
  set to three hours assuming \code{x} given in seconds.
  You can detect a better bandwidth by calling
  \code{\link{bandwidthCV}}}

  \item{locfit.deg}{\code{numeric}: \code{deg} parameter
  (polynomials degree) in call to
  \code{\link[locfit:locfit]{locfit}}}

  \item{relative.height.at.lag}{Parameter used by
  \code{\link{guessCellGrowthParams}}}
}
\value{
  Fit as returned by \code{\link[locfit:locfit]{locfit}}
  for the "locfit" model and as returned by
  \code{\link{nls}} for the "logistic", "gompertz", "rosso"
  and "baranyi" models. The returned value also has an
  attribute \code{maxGrowthRate} valueing the inferred
  maximum growth rate as well as an attribute
  \code{pointOfMaxGrowthRate} valuing the datapoint at
  which the growth rate is maximal. Also, it has an
  attribute \code{max} valuing the inferred maximum among
  the time points as well as \code{pointOfMax} valuing the
  datapoint of max fitted value. It gets the additional
  class \code{cellCurveFit} assigned.
}
\description{
  Fit a cell growth curve
}
\details{
  For the non-parametric "locfit" model, local regression
  is done by a call to \code{\link[locfit:locfit]{locfit}}.
  The returned maximum growth rate values the maximum value
  of the fitted derivative over the data points. For the
  parametric models "logistic", "gompertz", "rosso" and
  "baranyi", the function does a non-least square fit by
  calling \code{\link{nls}}. Initial parameters values are
  generated by \code{\link{guessCellGrowthParams}}. The
  returned maximum growth rate values the \code{mu}
  parameter of these models.
}
\examples{
x = 1:1000
          z = gompertz(x, mu=0.01, l=200, z0=1, zmax=5) + rnorm(length(x),sd=0.1)
          f = fitCellGrowth(x, z, model = "gompertz")
          floc = fitCellGrowth(x, z, model = "locfit", locfit.h=500)
          	plot(x,z, main="simulated data\\nGompertz model")
          	lines(x, predict(f,x), lwd=2, col="red")
          	lines(x, predict(floc,x), lwd=2, col="blue")
          	legend( "right", legend=c("gompertz fit", "locfit"), lwd=1, col=c("red","blue") )
}
\author{
  Julien Gagneur and Moritz Matthey
}
\seealso{
  \code{\link{nls}}, \code{\link[locfit:locfit]{locfit}},
  \code{\link{baranyi}}, \code{\link{gompertz}},
  \code{\link{logistic}}, \code{\link{rosso}},
  \code{\link{guessCellGrowthParams}},
  \code{\link{fitCellGrowths}}
}

