\name{getWellIdsTecan}
\alias{getWellIdsTecan}
\title{Get aliases for wells}
\usage{
  getWellIdsTecan(data)
}
\arguments{
  \item{data}{parsed data of the file. See
  \code{\link{readYeastGrower}} and
  \code{\link{readGenios}}}
}
\value{
  vector containing the aliases
}
\description{
  The aliases are generated by extracting the information
  from parsed data of the file generated by the tecan
  machine. See \code{\link{readYeastGrower}} and
  \code{\link{readGenios}}.
}
\examples{
data <- readYeastGrower( system.file("extdata", "Plate1_YPFruc.txt", package="cellGrowth"))
ids <- getWellIdsTecan(data)
}
\author{
  Julien Gagneur, Andreas Neudecker
}

