\name{plotPlate}
\alias{plotPlate}
\title{Plot of a well plate}
\usage{
  plotPlate(file, labels = NULL,
    fileParser = readYeastGrower,
    getWellIds = getWellIdsTecan, calibration = identity,
    extractRowColumn = getRowColumn, cex = 0.05,
    scaleX = 1, scaleY = log2, strip.lines = 1.05,
    strip.cex = 0.8, xlab = "time", ylab = "log2(OD)",
    main = basename(file),
    scales = list(x = list(rot = 45)), ...)
}
\arguments{
  \item{file}{file name}

  \item{labels}{\code{vector} of characters indicating the
  label of the wells}

  \item{fileParser}{the file parser which reads the file
  generated by the machine}

  \item{getWellIds}{function or vector. If getWellIds is a
  function its parameter is the parsed data of the file
  parsed by \code{fileParser}. It should return a vector
  containing the well identifiers, e.g. A01, A02, .. You
  can as well set the well identifiers as a vector
  directly}

  \item{calibration}{calibration function applied to the
  raw data (before scaleY is applied)}

  \item{extractRowColumn}{function which converts well
  identifiers into row and column names}

  \item{cex}{plot parameter}

  \item{scaleX}{factor which scales the x-axis}

  \item{scaleY}{function how to convert the y-axis ( e.g.
  log2 )}

  \item{strip.lines}{height in lines of the labels}

  \item{strip.cex}{text-size of the labels}

  \item{xlab}{plot parameter}

  \item{ylab}{plot parameter}

  \item{main}{plot parameter}

  \item{scales}{plot parameter}

  \item{...}{optional plot parameter. See details}
}
\description{
  Plot of a well plate directly from a file using a lattice
  xyplot
}
\details{
  All plot parameters are passed to the
  \code{\link{xyplot}} function
}
\examples{
plotPlate( system.file("extdata", "tecan_genios.txt", package="cellGrowth"), fileParser=readGenios)
}
\author{
  Andreas Neudecker
}

